Moves fetch, get campaign functions into helper
