+++ lib/helpers.js
@@ -6,6 +6,12 @@
 const logger = require('winston');
 const newrelic = require('newrelic');
 const stathat = require('./stathat');
+const helpers = require('./helpers/index');
+
+// register helpers
+Object.keys(helpers).forEach((helperName) => {
+  module.exports[helperName] = helpers[helperName];
+});
 
 /**
  * addBlinkSuppressHeaders
