+++ lib/helpers/campaign.js
@@ -2,10 +2,41 @@
 
 const dateFns = require('date-fns');
 const logger = require('winston');
+const Promise = require('bluebird');
+const cacheHelper = require('./cache');
+const phoenix = require('../phoenix');
 const config = require('../../config/lib/helpers/campaign');
 const phoenixConfig = require('../../config/lib/phoenix');
 
 module.exports = {
+  fetchById: function fetchById(campaignId) {
+    return new Promise((resolve, reject) => {
+      phoenix.fetchCampaignById(campaignId)
+        .then((res) => {
+          const campaignData = this.parseCampaign(res.data);
+          return cacheHelper.campaigns.set(`${campaignId}`, campaignData);
+        })
+        .then(campaign => resolve(JSON.parse(campaign)))
+        .catch(err => reject(err));
+    });
+  },
+  fetchByIds: function fetchByIds(campaignIds) {
+    const promises = [];
+    campaignIds.forEach(campaignId => promises.push(this.getById(campaignId)));
+    return Promise.all(promises);
+  },
+  getById: function getById(campaignId) {
+    return cacheHelper.campaigns.get(`${campaignId}`)
+      .then((campaign) => {
+        if (campaign) {
+          logger.debug(`Campaigns cache hit:${campaignId}`);
+          return Promise.resolve(campaign);
+        }
+        logger.debug(`Campaigns cache miss:${campaignId}`);
+        return this.fetchById(campaignId);
+      })
+      .catch(err => Promise.reject(err));
+  },
   /**
    * @param {Object} campaign
    * @return {Boolean}
