+++ lib/middleware/campaigns-index/phoenix-campaigns.js
@@ -1,13 +1,12 @@
 'use strict';
 
-const phoenix = require('../../phoenix');
 const helpers = require('../../helpers');
 
 module.exports = function getPhoenixCampaigns() {
   return (req, res, next) => {
     const campaignIds = Object.keys(req.keywordsByCampaignId);
 
-    return phoenix.fetchCampaigns(campaignIds)
+    return helpers.campaign.fetchByIds(campaignIds)
       .then((phoenixRes) => {
         req.phoenixCampaigns = phoenixRes;
         return next();
