+++ lib/middleware/campaigns-single/contentful-keywords.js
@@ -5,7 +5,7 @@ const helpers = require('../../helpers');
 
 module.exports = function getKeywords() {
   return (req, res, next) => {
-    contentful.fetchKeywordsForCampaignId(req.campaign.id)
+    contentful.fetchKeywordsForCampaignId(req.campaignId)
       .then((contentfulRes) => {
         req.campaign.keywords = contentfulRes;
         return next();
