+++ lib/middleware/campaigns-single/phoenix-campaign.js
@@ -1,13 +1,12 @@
 'use strict';
 
-const phoenix = require('../../phoenix');
 const helpers = require('../../helpers');
 
 module.exports = function getPhoenixCampaign() {
   return (req, res, next) => {
     req.campaignId = req.params.campaignId;
 
-    return phoenix.getCampaignById(req.campaignId)
+    return helpers.campaign.getById(req.campaignId)
       .then((phoenixRes) => {
         req.campaign = phoenixRes;
         return next();
