+++ lib/phoenix.js
@@ -6,8 +6,6 @@
 const superagent = require('superagent');
 const logger = require('winston');
 const Promise = require('bluebird');
-const cacheHelper = require('./helpers/cache');
-const campaignHelper = require('./helpers/campaign');
 const config = require('../config/lib/phoenix');
 
 /**
@@ -16,14 +14,16 @@ const config = require('../config/lib/phoenix');
  * @param {object} data
  */
 function executeGet(endpoint, query = {}) {
-  const baseUri = config.clientOptions.baseUri;
-  const url = `${baseUri}/${endpoint}`;
-  logger.debug('executeGet', { url });
+  const url = `${config.clientOptions.baseUri}/${endpoint}`;
 
-  return superagent
-    .get(url)
-    .query(query)
-    .then(res => res.body);
+  return new Promise((resolve, reject) => {
+    logger.debug('executeGet', { url });
+    return superagent
+      .get(url)
+      .query(query)
+      .then(res => resolve(res.body))
+      .catch(err => reject(module.exports.parsePhoenixError(err)));
+  });
 }
 
 /**
@@ -33,7 +33,6 @@ function executeGet(endpoint, query = {}) {
 module.exports.parsePhoenixError = function (err, campaignId) {
   const scope = err;
   scope.message = `phoenix campaignId=${campaignId} error=${err.message}`;
-
   return scope;
 };
 
@@ -57,48 +56,5 @@ module.exports.isClosedCampaign = function (phoenixCampaign) {
 module.exports.fetchCampaignById = function (campaignId) {
   logger.debug(`phoenix.fetchCampaignById:${campaignId}`);
   const endpoint = `campaigns/${campaignId}`;
-
-  return new Promise((resolve, reject) => {
-    executeGet(endpoint)
-      .then((res) => {
-        const campaignData = campaignHelper.parseCampaign(res.data);
-        return cacheHelper.campaigns.set(`${campaignId}`, campaignData);
-      })
-      .then(campaign => resolve(JSON.parse(campaign)))
-      .catch(err => reject(exports.parsePhoenixError(err, campaignId)));
-  });
-};
-
-/**
- * getCampaignById - Attemps to get the cached campaign by id, if its not
- *                   cached, it fetches it from Phoenix.
- *
- * @param  {Number} campaignId
- * @return {Promise}
- */
-module.exports.getCampaignById = function getCampaignById(campaignId) {
-  logger.debug(`phoenix.getCampaignById:${campaignId}`);
-
-  return cacheHelper.campaigns.get(`${campaignId}`)
-    .then((campaign) => {
-      if (campaign) {
-        logger.debug(`Campaigns cache hit:${campaignId}`);
-        return Promise.resolve(campaign);
-      }
-      logger.debug(`Campaigns cache miss:${campaignId}`);
-      return exports.fetchCampaignById(campaignId);
-    })
-    .catch(err => Promise.reject(err));
-};
-
-/**
- * @param {Array} campaignIds
- * @return {Promise}
- */
-module.exports.fetchCampaigns = function (campaignIds) {
-  const promises = [];
-  campaignIds.forEach((campaignId) => {
-    promises.push(module.exports.getCampaignById(campaignId));
-  });
-  return Promise.all(promises);
+  return executeGet(endpoint);
 };
